Rest Service with IoC
- Some classes that use RPC-XML calls have already been deprecated.
- Rest Service with IoC
