+++ livingdoc-confluence-plugin/src/main/java/info/novatec/testit/livingdoc/server/rpc/xmlrpc/LivingDocXmlRpcServer.java
@@ -67,7 +67,12 @@ import info.novatec.testit.livingdoc.server.transfer.SpecificationLocation;
  * Copyright (c) 2006 Pyxis technologies inc. All Rights Reserved.
  *
  * @author jchuet
+ *
+ * @deprecated The XML-RPC and SOAP APIs are deprecated since Confluence 5.5.
+ * More info <a href="https://developer.atlassian.com/confdev/deprecated-apis/confluence-xml-rpc-and-soap-apis">here</a>
+ * <br> Use {@link info.novatec.testit.livingdoc.confluence.rest.LivingDocRestService} instead.
  */
+@Deprecated
 public class LivingDocXmlRpcServer implements RpcServerService {
     private static final Logger log = LoggerFactory.getLogger(LivingDocXmlRpcServer.class);
     public static final String NUMBER = " number: ";
